+++ src/esphome/entities/light.lua
@@ -24,7 +24,14 @@ end
 function LightEntity:discovered(entity)
   log:trace("LightEntity:discovered(%s)", entity)
   local bindingId = assert(
-    bindings:getOrAddDynamicBinding(self.TYPE, "light_" .. entity.key, "PROXY", true, entity.name, "ESPHOME_LIGHT")
+    bindings:getOrAddDynamicBinding(
+      self.TYPE,
+      "light_" .. ESPHomeClient.entityRef(entity),
+      "PROXY",
+      true,
+      entity.name,
+      "ESPHOME_LIGHT"
+    )
   ).bindingId
   RFP[bindingId] = function(idBinding, strCommand, tParams, args)
     log:trace("RFP idBinding=%s strCommand=%s tParams=%s args=%s", idBinding, strCommand, tParams, args)
@@ -74,7 +81,7 @@ end
 --- @return void
 function LightEntity:updated(entity, state)
   log:trace("LightEntity:updated(%s, %s)", entity, state)
-  local binding = bindings:getDynamicBinding(self.TYPE, "light_" .. entity.key)
+  local binding = bindings:getDynamicBinding(self.TYPE, "light_" .. ESPHomeClient.entityRef(entity))
   if binding ~= nil then
     SendToProxy(binding.bindingId, "UPDATE_STATE", {
       entity = SerializeSafe(entity),